backend/requirements.txt
Flask==3.1.1
feedparser==6.0.11
APScheduler==3.11.0
SQLAlchemy==2.0.41
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0 # Added for database migrations
python-dateutil
gunicorn
Flask-Caching>=2.1.0
redis
Flask-Bcrypt
PyJWT
Flask-HTTPAuth
